[FIX] change statement term to statements
