--- l10n_it_financial_statement_eu/report/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2022 Associazione Odoo Italia (<http://www.odoo-italia.org>)
-# Copyright 2022 MKT Srl (<www.mkt.it>)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import financial_statement_eu_report
