--- l10n_it_financial_statement_eu/security/financial_statement_eu_security.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2022 MKT srl
-     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
-<odoo>
-    <record id="financial_statement_eu_access" model="ir.model.access">
-        <field name="name">financial.statement.eu.user</field>
-        <field name="model_id" ref="model_financial_statement_eu" />
-        <field name="group_id" ref="base.group_user" />
-        <field name="perm_read" eval="1" />
-        <field name="perm_create" eval="0" />
-        <field name="perm_write" eval="1" />
-        <field name="perm_unlink" eval="0" />
-    </record>
-
-    <record id="financial_statement_eu_wizard_access" model="ir.model.access">
-        <field name="name">financial.statement.eu.wizard.user</field>
-        <field name="model_id" ref="model_financial_statement_eu_wizard" />
-        <field name="group_id" ref="base.group_user" />
-        <field name="perm_read" eval="1" />
-        <field name="perm_create" eval="1" />
-        <field name="perm_write" eval="1" />
-        <field name="perm_unlink" eval="1" />
-    </record>
-
-    <record id="financial_statement_eu_log_access" model="ir.model.access">
-        <field name="name">financial.statement.eu.log.user</field>
-        <field name="model_id" ref="model_financial_statement_eu_log" />
-        <field name="group_id" ref="base.group_user" />
-        <field name="perm_read" eval="1" />
-        <field name="perm_create" eval="1" />
-        <field name="perm_write" eval="1" />
-        <field name="perm_unlink" eval="1" />
-    </record>
-
-
-</odoo>
