--- l10n_it_financial_statement_eu/static/src/css/report.css
@@ -1,118 +1,0 @@
-.act_as_table {
-    display: table !important;
-    background-color: white;
-}
-.act_as_row {
-    display: table-row !important;
-    page-break-inside: avoid;
-}
-.act_as_cell {
-    display: table-cell !important;
-    page-break-inside: avoid;
-}
-.act_as_thead {
-    display: table-header-group !important;
-}
-.act_as_tbody {
-    display: table-row-group !important;
-}
-.list_table,
-.data_table,
-.totals_table {
-    width: 100% !important;
-}
-.act_as_row.labels {
-    background-color: #f0f0f0 !important;
-}
-.list_table,
-.data_table,
-.totals_table,
-.list_table .act_as_row {
-    border-left: 0px;
-    border-right: 0px;
-    text-align: center;
-    font-size: 10px;
-    padding-right: 3px;
-    padding-left: 3px;
-    padding-top: 2px;
-    padding-bottom: 2px;
-    border-collapse: collapse;
-}
-.totals_table {
-    font-weight: bold;
-    text-align: center;
-}
-.list_table .act_as_row.labels,
-.list_table .act_as_row.initial_balance,
-.list_table .act_as_row.lines {
-    border-color: grey !important;
-    border-bottom: 1px solid lightGrey !important;
-}
-.data_table .act_as_cell {
-    border: 1px solid lightGrey;
-    text-align: center;
-}
-.data_table .act_as_cell,
-.list_table .act_as_cell,
-.totals_table .act_as_cell {
-    word-wrap: break-word;
-}
-.data_table .act_as_row.labels,
-.totals_table .act_as_row.labels {
-    font-weight: bold;
-}
-.initial_balance .act_as_cell {
-    font-style: italic;
-}
-.account_row {
-    font-size: 0.8em;
-    font-style: italic;
-}
-.account_row_background {
-    background-color: #efefef;
-}
-.period {
-    padding-bottom: 20px;
-}
-.act_as_cell.amount {
-    word-wrap: normal;
-    text-align: right;
-}
-.act_as_cell.left {
-    text-align: left;
-}
-.act_as_cell.right {
-    text-align: right;
-}
-.list_table .act_as_cell.first_column {
-    padding-left: 0px;
-}
-.overflow_ellipsis {
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-}
-.custom_footer {
-    font-size: 7px !important;
-}
-.page_break {
-    page-break-inside: avoid;
-}
-
-.button_row {
-    padding-bottom: 10px;
-}
-
-.o_financial_statement_eu_page {
-    padding-top: 10px;
-    width: 90%;
-    margin-right: auto;
-    margin-left: auto;
-    font-family: Helvetica, Arial;
-}
-
-.warnings {
-    padding: 10px;
-    color: #af0000;
-    background-color: #ffdfdf;
-}
