--- l10n_it_financial_statement_eu/static/src/xml/report.xml
@@ -1,38 +1,0 @@
-<template>
-    <!-- Buttons of the Control Panel -->
-    <t
-        t-name="info.ReportAction"
-        t-inherit="web.ReportAction"
-        t-inherit-mode="extension"
-        owl="1"
-    >
-        <xpath expr="//button" position="after">
-<!--            <button-->
-<!--                t-on-click="on_click_xlsx"-->
-<!--                type="button"-->
-<!--                class="btn btn-secondary"-->
-<!--                title="Export XLSX"-->
-<!--            >Export XLSX</button>-->
-<!--            <button-->
-<!--                t-on-click="on_click_xlsx"-->
-<!--                type="button"-->
-<!--                class="btn btn-secondary"-->
-<!--                title="Export XLSX"-->
-<!--            >Export XBRL</button>-->
-            <button
-                t-if="isFinancialStatementEU"
-                type="button"
-                t-on-click="export_xlsx"
-                class="btn btn-secondary o_report_xlsx"
-                title="Export XLSX"
-            >Export XLSX</button>
-            <button
-                t-if="isFinancialStatementEU"
-                type="button"
-                t-on-click="export_xbrl"
-                class="btn btn-secondary o_report_xbrl"
-                title="Export XBRL"
-            >Export XBRL</button>
-        </xpath>
-    </t>
-</template>
