--- l10n_it_financial_statement_eu/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_financial_statement_eu
