+++ l10n_it_financial_statements_eu/models/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2022 Associazione Odoo Italia (<http://www.odoo-italia.org>)
+# Copyright 2022 MKT Srl (<www.mkt.it>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import financial_statements_eu
