+++ l10n_it_financial_statements_eu/readme/CONTRIBUTORS.rst
