+++ l10n_it_financial_statements_eu/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_financial_statements_eu
