--- setup/l10n_it_financial_statement_eu/odoo/addons/l10n_it_financial_statement_eu
@@ -1,1 +1,0 @@
-../../../../l10n_it_financial_statement_eu
