+++ setup/l10n_it_financial_statements_eu/odoo/addons/l10n_it_financial_statements_eu
@@ -1,0 +1,1 @@
+../../../../l10n_it_financial_statements_eu
